✅ Rewrite tests on next generation of floating point arbitraries
